+++ src/Credfeto.DotNet.Repo.Tools.Packages/ISinglePackageUpdater.cs
@@ -18,6 +18,7 @@ public interface ISinglePackageUpdater
         BuildSettings buildSettings,
         DotNetVersionSettings dotNetSettings,
         PackageUpdate package,
+        string? knownTrackingHash,
         CancellationToken cancellationToken
     );
 }